requirements.txt
appdirs
APScheduler
astroid
autopep8
bs4
CacheControl
cachetools
certifi
cffi
chardet
colorama
contextlib2
cryptography>=41.0.5
decorator
distlib
distro
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
html5lib
httplib2
hurry.filesize
idna
ipaddr
isort
lazy-object-proxy
lockfile
mccabe
msgpack
oauth2client
packaging
pep517
progress
protobuf
pyasn1
pyasn1-modules
pycodestyle
pycparser
PyDrive
PyDrive2
pylint
pyOpenSSL>=24.0.0
pyparsing
python-dotenv
python-telegram-bot==13.0
pytoml
pytz
PyYAML
requests
retrying
rope
rsa
six
toml
tornado
tqdm
tzlocal
uritemplate
urllib3
webencodings
wrapt
yt-dlp
